requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Message Brokers
aiokafka==0.10.0
paho-mqtt==1.6.1

# ML & Analytics
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
